+++ ta1/NodeMulop.java
@@ -7,7 +7,7 @@ public class NodeMulop extends Node {
 		this.mulop=mulop;
 	}
 
-	public int op(int o1, int o2) throws EvalException {
+	public double op(double o1, double o2) throws EvalException {
 		if (mulop.equals("*"))
 			return o1*o2;
 		if (mulop.equals("/"))
